+++ solstice/CHANGELOG.md
@@ -5,6 +5,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 ## [Unreleased]
 ### Added
 - `WorkspaceRegistry` now supports cleaning. ([#50](https://github.com/equodev/equo-ide/pull/50))
+- Introduce `P2Model` so that Gradle and Maven DSLs can drive guts through a common model type. ([#51](https://github.com/equodev/equo-ide/pull/51))
 
 ## [0.9.0] - 2023-01-09
 ### Added